Split updaters into many files in a subfolder
