+++ rundmcmc/defaults.py
@@ -12,9 +12,7 @@ from rundmcmc.proposals import \
     propose_random_flip_no_loops as propose_random_flip
 from rundmcmc.updaters import (Tally, boundary_nodes, county_splits, cut_edges,
                                cut_edges_by_part, exterior_boundaries,
-                               perimeters)
-from rundmcmc.updaters import polsby_popper_updater as polsby_popper
-from rundmcmc.updaters import votes_updaters
+                               perimeters, polsby_popper, votes_updaters)
 from rundmcmc.validity import (L1_reciprocal_polsby_popper, UpperBound,
                                Validator, no_vanishing_districts,
                                refuse_new_splits, single_flip_contiguous,
