+++ tests/test_defaults.py
@@ -1,0 +1,5 @@
+import rundmcmc.defaults
+
+
+def test_defaults_are_importable():
+    assert rundmcmc.defaults
